RunJavascript and basic edit/save/switch page working
